Kickstarter_Challenge
Louise is planning a theater play “Fever”. The estimated budget is over $10,000 and she already came close to its fundraising goal in a short amount of time. 

The purpose of the analysis is to determine whether there are specific factors that makes a crowdfunding campaign successful and compare how different campaigns fared in relation to their launch dates and their funding goals.